requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

attr==0.3.2
azure-functions
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.3
azure-storage-blob
dl==0.1.0
docutils==0.20.1
ipython==8.15.0
ipywidgets==8.1.1
Jinja2==3.1.2
keyring==24.2.0
lockfile==0.12.2
matplotlib==3.8.0
mock==5.1.0
numpy==1.26.0
pandas==2.1.1
protobuf==4.24.3


railroad==0.5.0
scikit_learn==1.3.1
scipy==1.11.2
Sphinx==7.2.6

trove_classifiers==2023.9.19

